Customization of header.html
To override the default structure and style of minima, simply create
the concerned directory at the root of your site, copy the file you
wish to customize to the directory, and then edit the file e.g. to
override the '_includes/head.html' file to specify a custom style
path, create an _includes directory, copy _includes/head.html from
the minima gem folder to <yoursite>/_includes and start editing the
file.

Favicon
Created an image called notapipe.png and processed it in
realfavicongenerator.net that includes the files and the text
to add to head.html.

Once that was added to the project it 'just worked'